Only download Git LFS when GIT_LFS_VERSION is set

+++ script/build-arm64.sh
@@ -19,26 +19,32 @@ docker run -it \
 --rm multiarch/debian-debootstrap:arm64-jessie sh $BASEDIR/script/build-arm64-git.sh
 cd - > /dev/null
 
-echo "-- Building Git LFS"
-go get github.com/git-lfs/git-lfs
-GOPATH=`go env GOPATH`
-cd $GOPATH/src/github.com/git-lfs/git-lfs
-git checkout "v${GIT_LFS_VERSION}"
-# Make the 'mangen' target first, without setting GOOS/GOARCH.
-make mangen
-make GOARCH=arm64 GOOS=linux
-GIT_LFS_OUTPUT_DIR=$GOPATH/src/github.com/git-lfs/git-lfs/bin/
 
-echo "-- Verifying built Git LFS"
-docker run -it \
- --mount type=bind,source=$GIT_LFS_OUTPUT_DIR,target=$GIT_LFS_OUTPUT_DIR \
- -w=$BASEDIR \
- --rm multiarch/debian-debootstrap:arm64-jessie $GIT_LFS_OUTPUT_DIR/git-lfs --version
+if [[ "$GIT_LFS_VERSION" ]]; then
+  echo "-- Building Git LFS"
+  go get github.com/git-lfs/git-lfs
+  GOPATH=`go env GOPATH`
+  cd $GOPATH/src/github.com/git-lfs/git-lfs
+  git checkout "v${GIT_LFS_VERSION}"
+  # Make the 'mangen' target first, without setting GOOS/GOARCH.
+  make mangen
+  make GOARCH=arm64 GOOS=linux
+  GIT_LFS_OUTPUT_DIR=$GOPATH/src/github.com/git-lfs/git-lfs/bin/
 
-echo "-- Bundling Git LFS"
-GIT_LFS_FILE=$GIT_LFS_OUTPUT_DIR/git-lfs
-SUBFOLDER="$DESTINATION/libexec/git-core"
-cp $GIT_LFS_FILE $SUBFOLDER
+  echo "-- Verifying built Git LFS"
+  docker run -it \
+   --mount type=bind,source=$GIT_LFS_OUTPUT_DIR,target=$GIT_LFS_OUTPUT_DIR \
+   -w=$BASEDIR \
+   --rm multiarch/debian-debootstrap:arm64-jessie $GIT_LFS_OUTPUT_DIR/git-lfs --version
+
+  echo "-- Bundling Git LFS"
+  GIT_LFS_FILE=$GIT_LFS_OUTPUT_DIR/git-lfs
+  SUBFOLDER="$DESTINATION/libexec/git-core"
+  cp $GIT_LFS_FILE $SUBFOLDER
+else
+  echo "-- Skipping Git LFS"
+fi
+
 
 # download CA bundle and write straight to temp folder
 # for more information: https://curl.haxx.se/docs/caextract.html
